pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "traffic-similarity-analysis"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "dash>=2.17",
    "dash-bootstrap-components>=1.6",
    "plotly>=5.22",
    "pandas>=2.2",
    "numpy>=1.26",
    "scipy>=1.13",
    "scikit-learn>=1.5",
    "lightgbm>=4.3",
    "shap>=0.45",
    "umap-learn>=0.5",
    "openpyxl>=3.1",
]

[project.optional-dependencies]
dev = ["pytest>=8", "pytest-cov"]

[tool.pytest.ini_options]
testpaths = ["tests"]
